+++ sentry-android-core/src/test/java/io/sentry/android/core/DefaultAndroidEventProcessorTest.kt
@@ -27,6 +27,7 @@ import io.sentry.protocol.SentryTransaction
 import io.sentry.protocol.User
 import io.sentry.test.getCtor
 import org.junit.runner.RunWith
+import java.util.*
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -374,6 +375,18 @@ class DefaultAndroidEventProcessorTest {
 //            assertNotNull(device.externalFreeStorage)
 //            assertNotNull(device.externalStorageSize)
 //            assertNotNull(device.connectionType)
+        }
+    }
+
+    @Test
+    fun `Event sets language and locale`() {
+        val sut = fixture.getSut(context)
+        val locale = Locale.getDefault()
+
+        assertNotNull(sut.process(SentryEvent(), null)) {
+            val device = it.contexts.device!!
+            assertEquals(device.language, locale.language)
+            assertEquals(device.locale, locale.toString())
         }
     }
 }