+++ sentry/src/main/java/io/sentry/protocol/Device.java
@@ -93,7 +93,18 @@ public final class Device implements IUnknownPropertiesConsumer {
   private @Nullable TimeZone timezone;
 
   private @Nullable String id;
-  private @Nullable String language;
+
+  /**
+   * This method returns the language code for this locale, which will either be the empty string or
+   * a lowercase ISO 639 code.
+   *
+   * @deprecated use {@link Device#getLocale()}
+   */
+  @Deprecated private @Nullable String language;
+
+  /** The locale of the device. For example, en-US. */
+  private @Nullable String locale;
+
   private @Nullable String connectionType;
 
   /** battery's temperature in celsius */
@@ -135,6 +146,7 @@ public final class Device implements IUnknownPropertiesConsumer {
     this.batteryLevel = device.batteryLevel;
     final String[] archsRef = device.archs;
     this.archs = archsRef != null ? archsRef.clone() : null;
+    this.setLocale(device.getLocale());
 
     final TimeZone timezoneRef = device.timezone;
     this.timezone = timezoneRef != null ? (TimeZone) timezoneRef.clone() : null;
@@ -382,6 +394,14 @@ public final class Device implements IUnknownPropertiesConsumer {
 
   public void setBatteryTemperature(final @Nullable Float batteryTemperature) {
     this.batteryTemperature = batteryTemperature;
+  }
+
+  public @Nullable String getLocale() {
+    return locale;
+  }
+
+  public void setLocale(final @Nullable String locale) {
+    this.locale = locale;
   }
 
   @TestOnly