+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Ref: Rename Fragment span operation from `ui.fragment.load` to `ui.load` (#1824)
+* Feat: Add locale to device context and deprecate language (#)
 
 ## 5.4.3